+++ database/migrations/2017_08_28_211622_create_azucar_table.php
@@ -17,7 +17,7 @@ class CreateAzucarTable extends Migration
           $table->increments('id');
           $table->string('DES_AZU');
           $table->string('EST_AZU');
-          $table->money('PRE_AZU');
+          $table->float('PRE_AZU');
           $table->timestamps();
       });
     }
